A few changes (SEE DESCRIPTION)
Added a grid, so the obstacle checkers no longer need to iterate through
all of the obstacles. Lots of debugging took place for that. And I added
a function to identify your current position. Also I have
partially-completed a function to auto-solve the maze. It's nowhere near
completion. But the maze works.